+++ packages/browser-wallet-api/src/wallet-api.ts
@@ -19,12 +19,14 @@ import {
     EventType,
     SchemaWithContext,
     SchemaType,
+    SignMessageObject,
     SmartContractParameters,
 } from '@concordium/browser-wallet-api-helpers';
 import EventEmitter from 'events';
 import type { JsonRpcRequest } from '@concordium/common-sdk/lib/providers/provider';
+import { IdProofOutput, IdStatement } from '@concordium/common-sdk/lib/idProofTypes';
+import ConcordiumGRPCClient from '@concordium/common-sdk/lib/GRPCClient';
 import JSONBig from 'json-bigint';
-import { ConcordiumGRPCClient, IdProofOutput, IdStatement } from '@concordium/common-sdk';
 import { stringify } from './util';
 import { BWGRPCTransport } from './gRPC-transport';
 
@@ -60,7 +62,10 @@ class WalletApi extends EventEmitter implements IWalletApi {
     /**
      * Sends a sign request to the Concordium Wallet and awaits the users action
      */
-    public async signMessage(accountAddress: string, message: string): Promise<AccountTransactionSignature> {
+    public async signMessage(
+        accountAddress: string,
+        message: string | SignMessageObject
+    ): Promise<AccountTransactionSignature> {
         const response = await this.messageHandler.sendMessage<MessageStatusWrapper<AccountTransactionSignature>>(
             MessageType.SignMessage,
             {
@@ -70,7 +75,7 @@ class WalletApi extends EventEmitter implements IWalletApi {
         );
 
         if (!response.success) {
-            throw new Error('Signing rejected');
+            throw new Error(response.message);
         }
 
         return response.result;