+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 2.4.0
+## Unreleased
 
 ### Added
 
@@ -10,6 +10,12 @@
 ### Deprecated
 
 -   `getJsonRpcClient` in favor of the new `getGrpcClient`.
+
+## 2.4.0
+
+### Added
+
+-   signMessage is now able to receive a SignMessageObject as the message in addition to a utf8 string.
 
 ## 2.3.0